requirements.txt
Flask==3.0.3
supabase==2.4.0
cloudinary==1.41.0
python-dotenv==1.0.1
